rename stac_io module to stac_api_io
