+++ pystac_client/item_search.py
@@ -11,7 +11,7 @@ import warnings
 from pystac import Collection, Item, ItemCollection, Link
 from pystac.stac_io import StacIO
 
-from pystac_client.stac_io import StacApiIO
+from pystac_client.stac_api_io import StacApiIO
 from pystac_client.conformance import ConformanceClasses, ConformanceMixin
 
 if TYPE_CHECKING:
@@ -181,7 +181,7 @@ class ItemSearch(ConformanceMixin):
         if self._max_items is not None and limit is not None:
             limit = min(limit, self._max_items)
 
-        if limit < 1 or limit > 10000:
+        if limit is not None and (limit < 1 or limit > 10000):
             raise Exception(f"Invalid limit of {limit}, must be between 1 and 10,000")
 
         self.method = method
