--- tests/cassettes/test_stac_io/TestSTAC_IOOverride.test_str_input.yaml
@@ -1,43 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.25.1
-    method: GET
-    uri: https://planetarycomputer.microsoft.com/api/stac/v1
-  response:
-    body:
-      string: '{"id":"microsoft-pc","description":"Searchable spatiotemporal metadata
-        describing Earth science datasets hosted by the Microsoft Planetary Computer","title":"Microsoft
-        Planetary Computer STAC API","stac_version":"1.0.0-beta.2","links":[{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/","rel":"self","type":"application/json"},{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/docs","rel":"docs","type":"text/html","title":"OpenAPI
-        docs"},{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/conformance","rel":"conformance","type":"application/json","title":"STAC/WFS3
-        conformance classes implemented by this server"},{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/search","rel":"search","type":"application/geo+json","title":"STAC
-        search"},{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/collections/aster-l1t","rel":"child","type":"application/json","title":"ASTER
-        L1T"},{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/collections/landsat-8-c2-l2","rel":"child","type":"application/json","title":"Landsat
-        8 Collection 2 Level-2"},{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/collections/sentinel-2-l2a","rel":"child","type":"application/json","title":"Sentinel-2
-        Level-2A"},{"href":"https://planetarycomputer.microsoft.com/api/stac/v1/collections/naip","rel":"child","type":"application/json","title":"NAIP:
-        National Agriculture Imagery Program"}],"conformsTo":["https://api.stacspec.org/v1.0.0-beta.1/core","https://api.stacspec.org/v1.0.0-beta.1/item-search","http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/core","http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas30","http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/geojson"]}'
-    headers:
-      Content-Length:
-      - '1742'
-      Content-Type:
-      - application/json
-      Date:
-      - Wed, 16 Jun 2021 02:54:00 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      X-Azure-Ref:
-      - 0yGfJYAAAAAA4529cCs+RRq4zsfhVRfc0Qk9TMzFFREdFMDQxMQA5MjdhYmZhNi0xOWY2LTRhZjEtYTA5ZC1jOTU5ZDlhMWU2NDQ=
-      X-Cache:
-      - CONFIG_NOCACHE
-    status:
-      code: 200
-      message: OK
-version: 1
